Compilation on Windows: use Cython, move to Setuptools
distutils is old and dead (long ago old and dead) and there needs to b
a change to something newer. This change comes in the form of moving
to setuptools, a cleaner, newer way of handling things. As it stands, we
can just require Cython (since we'd already need a compiler anyhow)
and deal with not having something checked in.

At release, we *should* check in a copy so that we don't *need* Cython
but it's useful to have and likely already installed. By adding it into
the `install_requires` group, we're more likely to have a copy anyhow.

We also should pass in the `/std:c++14` flag into MSVC, so we do now.