Travis: use pip for installation, include Cythonized version.
We should use Pip for our shenanigans. It calls setup.py and friends.

Travis needs pip to have the packages listed in requirements.txt
installed when you run nosetests

This isn't listed in setup.cfg or setup.py because setuptools figures
all that out for you *except* for nose since it's not supported
anymore. Moving to pytest might help that later on.

nosetests needs to be called directly, plus Travis needs Xenial
to use 3.7. Travis on Xenial and Travis on Trusty have two
different sets of semi-overlapping support sets for Python.
I'm serious. This sucks.

This also includes a Cythonized version of the package.